Move exceptions used in UI only from `pyramid` to `open` package
