+++ src/main/java/sc/fiji/ome/zarr/open/exceptions/NonExistingResolutionLevelException.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,18 +26,24 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.pyramid.exceptions;
+package sc.fiji.ome.zarr.open.exceptions;
 
-public class NotASingleScaleImageException extends RuntimeException
+public class NonExistingResolutionLevelException extends RuntimeException
 {
-
-	public NotASingleScaleImageException( final String path )
+	public NonExistingResolutionLevelException( final int resolutionLevel, final int numResolutionLevels )
 	{
-		super( "The dataset at path '" + path + "' is not a valid OME-Zarr single scale image." );
+		super( message( resolutionLevel, numResolutionLevels ) );
 	}
 
-	public NotASingleScaleImageException( final String path, final Throwable cause )
+	public NonExistingResolutionLevelException( final int resolutionLevel, final int numResolutionLevels, final Throwable cause )
 	{
-		super( "The dataset at path '" + path + "' is not a valid OME-Zarr single scale image. Cause: " + cause.getMessage(), cause );
+		super( message( resolutionLevel, numResolutionLevels ), cause );
+	}
+
+	private static String message( final int resolutionLevel, final int numResolutionLevels )
+	{
+		return "Resolution level " + resolutionLevel + " does not exist. "
+				+ "The pyramid has " + numResolutionLevels + " level"
+				+ ( numResolutionLevels == 1 ? "" : "s" ) + " (0–" + ( numResolutionLevels - 1 ) + ").";
 	}
 }
